recruitment_orchestration_platform/recruitment_platform/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.15
langgraph==0.2.45
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
PyPDF2==3.0.1
pypdf==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
httpx==0.27.2
faker==30.8.1
rich==13.9.2
loguru==0.7.2
